WF MatchHandler - New file added
MatchHandler is a versatile object that creates and wraps around TexasHoldEm objects. PokerBot child class instances are passed in at instantiation, and intuitive methods provide control flow for matches. Matches can be ran one round at a time, or one match at a time, or even multiple matches back to back.

PokerBot child class instances are passed by reference so that they can be trained in one MatchHandler instance, then transferred to and evaluated within another instance. This should provide us with the modularity we need for expanding our project in an intuitive package.

I intend for this class to be able to record match data in the future.