WF Many Files - Modified Control Flow
Remedying vocabulary mistakes and simplifying control flow:

In the past, we've been referring to what is actually a "hand" as a "round", although a "round" is an entirely different thing. Control flow flags have been updated with new names to remedy this.

Also, to make bots more autonomous, PokerBot and all current subclasses now receive flags with no arguments. This should prevent these flags from becoming too ridiculously niche, and only purposeful for a single bot.

I want to outsource as much complexity as possible to each individual bot to keep the project modular, and these changes should facilitate that.