WF JerryBotRational - modified params get and set
The previous parameter getters and setters in JerryBotRational were developed to serve as a manual substitute for Pickling technology.

The current version allows for the getting and setting of any combination of parameters for a better user experience, now that we're using Pickle for saving and loading. This methodology should be used in future agents.