WF MatchHandler - history recording done
Now, MatchHandler objects are capable of recording the histories of the matches that they run. I decided to always record history when using a MatchHandler, as it's not intensive or obstructive.

Due to the object oriented approach, later work will need to be done in adding interpreter and print methods to these objects, but as of right now, everything should function properly.